+++ swift/Examples/TailscaleKitCLI/Package.swift
@@ -3,10 +3,8 @@ import PackageDescription
 
 let package = Package(
     name: "TailscaleKitCLI",
-    // Linux only: this depends on the root package's CTailscale binaryTarget, which
-    // only ships Linux triples today. See ../../../Package.swift.
     dependencies: [
-        .package(name: "TailscaleKit", path: "../../.."),
+        .package(name: "TailscaleKit", path: "../.."),
         .package(url: "https://github.com/apple/swift-argument-parser", from: "1.8.0"),
     ],
     targets: [
